move modal content to component
